face_recognition_v2/requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cmake==3.15.3
cycler==0.10.0
dlib==19.18.0
docutils==0.15.2
face_recognition
face-recognition-models
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
idna==2.8
joblib==0.14.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib
mtcnn==0.0.9
numpy==1.19.3
opencv-python
pandas
Pillow
protobuf==3.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
scikit-learn
scipy
six==1.12.0
sklearn==0.0
statistics==1.0.3.5
tensorboard
tensorflow-estimator
tensorflow-gpu
termcolor==1.1.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
